Added code to encrypt service account id as capability; split auth policy manager into a separate file
