+++ skydentity/policies/checker/gcp_authorization_policy.py
@@ -1,19 +1,14 @@
 import os
 import yaml
 
+from base64 import b64encode
 from enum import Enum
+from Crypto.Cipher import AES
 from dataclasses import dataclass
 from typing import Dict, List, Optional
 from flask import Request
 
-import firebase_admin
-from firebase_admin import credentials
-from firebase_admin import firestore
-from skydentity.policies.checker.resource_policy import CloudPolicy
-
-from skydentity.policies.managers.policy_manager import PolicyManager
 from skydentity.policies.checker.authorization_policy import AuthorizationPolicy
-from skydentity.policies.iam.gcp_service_account_manager import GCPServiceAccountManager
 
 @dataclass
 class RestrictedRole:
@@ -128,35 +123,9 @@ class GCPAuthorizationPolicy(AuthorizationPolicy):
                 else:
                     print(f"Request is unrecognized (gcp_authorization_policy.py): {request.url}, {request.method}")
                 return (None, False)
-
+    
     def create_service_account_with_roles(self):
         pass
-
-class GCPAuthorizationPolicyManager(PolicyManager):
-    def __init__(self, 
-                 credentials_path: str,
-                 firestore_policy_collection: str = 'authorization_policies'):
-        """
-        Initializes the GCP policy manager.
-        :param credentials_path: The path to the credentials file.
-        """
-        self._cred = credentials.Certificate(credentials_path)
-        self._app = firebase_admin.initialize_app(self._cred)
-        self._db = firestore.client()
-        self._firestore_policy_collection = firestore_policy_collection
-
-    def get_policy_dict(self, public_key: str) -> CloudPolicy | None:
-        """
-        Gets a policy from the cloud vendor.
-        :param public_key: The public key of the policy.
-        :return: The policy.
-        """
-        return self._db \
-            .collection(self._firestore_policy_collection) \
-            .document(public_key) \
-            .get() \
-            .to_dict()
-
 
 def main():
     print("HELLO")
